requirements.txt
discord.py==2.0.1
Flask==2.2.2
replit==3.2.4
requests==2.28.1
waitress==2.1.2
